backend/requirements.txt
# Python 3.12
Django==5.0.3
django_advanced_password_validation==1.2.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
python-dotenv==1.0.1
xhtml2pdf==0.2.16
